+++ src/test/java/com/epam/test/ht12/models/responses/createpet/PetModel.java
@@ -1,7 +1,8 @@
 package com.epam.test.ht12.models.responses.createpet;
 
-import java.util.List;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
 
 public class PetModel {
 
@@ -12,7 +13,7 @@ public class PetModel {
 	private String name;
 
 	@JsonProperty("id")
-	private int id;
+	private long id;
 
 	@JsonProperty("category")
 	private Category category;
@@ -39,11 +40,11 @@ public class PetModel {
 		return name;
 	}
 
-	public void setId(int id){
+	public void setId(long id){
 		this.id = id;
 	}
 
-	public int getId(){
+	public long getId(){
 		return id;
 	}
